requirements.txt
flask
flask-cors
python-dotenv
azure-cognitiveservices-speech
ffmpeg-python
pydub
gunicorn
reportlab>=4.0